backend/pyproject.toml
[tool.poetry]
name = "daylesford"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.4"
django-filter = "^2.4.0"
djangorestframework = "^3.12.4"
requests = "^2.25.1"
python-docx = "^0.8.11"
PyYAML = "^5.4.1"
drf-spectacular = "^0.17.2"
docxtpl = "^0.11.5"
docx2pdf = "^0.1.7"
docopt = "^0.6.2"
num2words = "^0.5.10"
gunicorn = "^20.1.0"
djangorestframework-simplejwt = "^4.7.1"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.9.1"
drf-writable-nested = "^0.6.3"
pyOpenSSL = "^20.0.1"
djangorestframework-camel-case = "^1.2.0"
model-bakery = "^1.3.2"
pytest = "^6.2.4"
pytest-django = "^4.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
